requirements.txt
requests
pandas
azure-storage-file-datalake
azure.identity